backend/requirements.txt
appdirs==1.4.4
apturl==0.5.2
attrs==21.2.0
bcrypt==3.2.0
beniget==0.4.1
black==25.11.0
blinker==1.9.0
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
cfgv==3.5.0
chardet==4.0.0
click==8.3.1
colorama==0.4.4
cpuset==1.6
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
distlib==0.4.0
distro==1.7.0
distro-info==1.1+ubuntu0.2
DockerHub-API @ git+https://github.com/Granulate/DockerHub-API.git@33106605c175af662209aa0b20fd7c67096b3211
duplicity==0.8.21
fasteners==0.14.1
filelock==3.20.0
flake8==7.3.0
Flask==3.1.2
flask-cors==6.0.2
fonttools==4.29.1
fs==2.4.12
furl==2.1.4
future==0.18.2
gast==0.5.2
gyp==0.1
httplib2==0.20.2
identify==2.6.15
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
jetson-stats==4.3.2
Jetson.GPIO==2.1.7
Jinja2==3.1.6
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==3.0.3
matplotlib==3.5.1
mccabe==0.7.0
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
mypy_extensions==1.1.0
mysql-connector-python==9.5.0
nodeenv==1.9.1
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
onboard==1.4.1
orderedmultidict==1.0.2
packaging==25.0
pandas==1.3.5
paramiko==2.9.3
pathspec==0.12.1
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.5.0
ply==3.11
pre_commit==4.5.0
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycodestyle==2.14.0
pycups==2.0.1
pyflakes==3.4.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
PyOpenGL==3.1.5
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-dbusmock==0.27.5
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.2.1
pythran==0.10.0
pytokens==0.3.0
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.3
requests==2.25.1
scipy==1.8.0
SecretStorage==3.3.1
six==1.16.0
smbus2==0.5.0
sympy==1.9
systemd-python==234
tabulate==0.9.0
tensorrt==10.3.0
tensorrt_dispatch==10.3.0
tensorrt_lean==10.3.0
termcolor==3.2.0
tomli==2.3.0
typing_extensions==4.15.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unicodedata2==14.0.0
urllib3==1.26.5
urwid==2.1.2
virtualenv==20.35.4
wadllib==1.3.6
Werkzeug==3.1.4
wget==3.2
xdg==5
xkit==0.0.0
zipp==1.0.0
